+++ crates/consensus/src/lib.rs
@@ -87,6 +87,7 @@ pub mod serde_bincode_compat {
 pub mod private {
     pub use alloy_eips;
     pub use alloy_primitives;
+    pub use alloy_rlp;
     #[cfg(feature = "arbitrary")]
     pub use arbitrary;
     #[cfg(feature = "serde")]
